Revert "Merge develop to master (#1791)"
This reverts commit 1c6aa89efecdbdfdb40f3b402d1f3af75a8116d1.

+++ src/main/scala/beam/utils/scenario/Models.scala
@@ -17,4 +17,4 @@ case class PlanElement(
   mode: Option[String]
 )
 
-case class HouseholdInfo(householdId: HouseholdId, cars: Int, income: Double, x: Double, y: Double)
+case class HouseholdInfo(householdId: HouseholdId, cars: Double, income: Double, x: Double, y: Double)
